+++ src/main/java/softeer2nd/chess/pieces/concrete/Bishop.java
@@ -1,4 +1,6 @@
-package softeer2nd.chess.pieces;
+package softeer2nd.chess.pieces.concrete;
+
+import softeer2nd.chess.pieces.RecursiveMovePiece;
 
 import java.util.List;
